+++ packages/mcp/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.0](https://github.com/microlinkhq/microlink/compare/%40microlink%2Fmcp%402.6.2...%40microlink%2Fmcp%402.7.0) (2026-09-17)
+
+### Features
+
+* **cli:** add buy command ([#59](https://github.com/microlinkhq/microlink/issues/59)) ([61ce1ec](https://github.com/microlinkhq/microlink/commit/61ce1ec519f4bb01717212d6a877f49bfb3fd834))
+
+## [2.6.2](https://github.com/microlinkhq/microlink/compare/%40microlink%2Fmcp%402.6.1...%40microlink%2Fmcp%402.6.2) (2026-09-17)
+
+**Note:** Version bump only for package @microlink/mcp
+
+## [2.6.1](https://github.com/microlinkhq/microlink/compare/%40microlink%2Fmcp%402.6.0...%40microlink%2Fmcp%402.6.1) (2026-09-16)
+
+**Note:** Version bump only for package @microlink/mcp
+
 ## [2.6.0](https://github.com/microlinkhq/microlink/compare/%40microlink%2Fmcp%402.5.2...%40microlink%2Fmcp%402.6.0) (2026-09-16)
 
 ### Features